requirements.txt
pillow==8.0.1
torch==1.7.0
tensorboardX==0.8.1
torchvision
matplotlib==3.3.3
numpy
scipy
opencv-python==3.4.2.16
opencv-contrib-python==3.4.2.16
tqdm
pyyaml
jupyter